# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pulse - Distributed Job Queue SaaS API",
    "version": "1.0.0",
    "description": "REST API for the Pulse distributed job queue and worker orchestration platform."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/organizations/{slug}/projects": {
      "post": {
        "tags": ["Projects"],
        "summary": "Create project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "createProject",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProjectRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Project created" } }
      },
      "get": {
        "tags": ["Projects"],
        "summary": "List projects",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "operationId": "listProjects",
        "responses": { "200": { "description": "Paginated projects" } }
      }
    },
    "/organizations/{slug}/projects/{projectSlug}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "getProject",
        "responses": { "200": { "description": "Project details" } }
      },
      "put": {
        "tags": ["Projects"],
        "summary": "Update project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "updateProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProjectRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Project updated" } }
      },
      "delete": {
        "tags": ["Projects"],
        "summary": "Delete project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "deleteProject",
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/organizations/{slug}/projects/{projectSlug}/queues": {
      "post": {
        "tags": ["Queues"],
        "summary": "Create queue",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "createQueue",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateQueueRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Queue created" } }
      },
      "get": {
        "tags": ["Queues"],
        "summary": "List queues",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "listQueues",
        "responses": { "200": { "description": "Paginated queues" } }
      }
    },
    "/organizations/{slug}/projects/{projectSlug}/queues/{name}": {
      "get": {
        "tags": ["Queues"],
        "summary": "Get queue",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "getQueue",
        "responses": { "200": { "description": "Queue details" } }
      },
      "put": {
        "tags": ["Queues"],
        "summary": "Update queue",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "updateQueue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateQueueRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Queue updated" } }
      },
      "delete": {
        "tags": ["Queues"],
        "summary": "Delete queue",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "deleteQueue",
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/organizations/{slug}/projects/{projectSlug}/queues/{name}/pause": {
      "post": {
        "tags": ["Queues"],
        "summary": "Pause queue",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "pauseQueue",
        "responses": { "200": { "description": "Queue paused" } }
      }
    },
    "/organizations/{slug}/projects/{projectSlug}/queues/{name}/resume": {
      "post": {
        "tags": ["Queues"],
        "summary": "Resume queue",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "resumeQueue",
        "responses": { "200": { "description": "Queue resumed" } }
      }
    },
    "/organizations/{slug}/projects/{projectSlug}/queues/{name}/stats": {
      "get": {
        "tags": ["Queues"],
        "summary": "Queue stats",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "getQueueStats",
        "responses": { "200": { "description": "Queue statistics" } }
      }
    },
...
    "/workers/register": {
      "post": {
        "tags": ["Workers"],
        "summary": "Register worker",
        "operationId": "registerWorker",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterWorkerRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Worker registered" } }
      }
    },
    "/workers/{id}/heartbeat": {
      "post": {
        "tags": ["Workers"],
        "summary": "Worker heartbeat",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "workerHeartbeat",
        "responses": { "200": { "description": "Heartbeat recorded" } }
      }
    },
    "/workers/{id}/claim": {
      "post": {
        "tags": ["Workers"],
        "summary": "Claim jobs",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "claimJobs",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ClaimJobsRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Jobs claimed" } }
      }
    },
    "/workers/{id}/jobs/{jobId}/complete": {
      "post": {
        "tags": ["Workers"],
        "summary": "Complete job",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "completeJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompleteJobRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Job completed" } }
      }
    },
    "/workers/{id}/jobs/{jobId}/fail": {
      "post": {
        "tags": ["Workers"],
        "summary": "Fail job",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "failJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FailJobRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Job failed" } }
      }
    },
...
    "/workers/{id}": {
      "get": {
        "tags": ["Workers"],
        "summary": "Get worker",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "getWorker",
        "responses": { "200": { "description": "Worker details" } }
      },
      "delete": {
        "tags": ["Workers"],
        "summary": "Delete worker",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "deleteWorker",
        "responses": { "204": { "description": "Deleted" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateProjectRequest": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string", "pattern": "^[a-z0-9-]+$" },
          "description": { "type": "string" }
        }
      },
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "CreateQueueRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 10 },
          "concurrencyLimit": { "type": "integer", "minimum": 1, "maximum": 100 },
          "retryStrategy": { "type": "string", "enum": ["FIXED", "LINEAR", "EXPONENTIAL"] },
          "maxRetries": { "type": "integer", "minimum": 0, "maximum": 100 },
          "baseRetryDelayMs": { "type": "integer", "minimum": 0, "maximum": 3600000 }
        }
      },
      "UpdateQueueRequest": {
        "type": "object",
        "properties": {
          "priority": { "type": "integer" },
          "concurrencyLimit": { "type": "integer" },
          "retryStrategy": { "type": "string" },
          "maxRetries": { "type": "integer" },
          "baseRetryDelayMs": { "type": "integer" }
        }
      },
...
      "RegisterWorkerRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "hostname": { "type": "string" },
          "concurrency": { "type": "integer", "minimum": 1, "maximum": 50 },
          "metadata": { "type": "object" }
        }
      },
      "ClaimJobsRequest": {
        "type": "object",
        "required": ["queueIds"],
        "properties": {
          "queueIds": { "type": "array", "items": { "type": "string" } },
          "limit": { "type": "integer", "minimum": 1 }
        }
      },
      "CompleteJobRequest": {
        "type": "object",
        "properties": {
          "result": { "type": "string" }
        }
      },
      "FailJobRequest": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}